+++ posthog-samples/posthog-android-sample/src/main/java/com/posthog/android/sample/MyApp.kt
@@ -2,8 +2,6 @@ package com.posthog.android.sample
 
 import android.app.Application
 import android.os.StrictMode
-import com.posthog.PostHogOnFeatureFlags
-import com.posthog.PostHogPropertiesSanitizer
 import com.posthog.android.PostHogAndroid
 import com.posthog.android.PostHogAndroidConfig
 
@@ -27,19 +25,19 @@ class MyApp : Application() {
                 captureDeepLinks = false
                 captureApplicationLifecycleEvents = false
                 captureScreenViews = false
-                sessionReplay = true
-                preloadFeatureFlags = true
-                onFeatureFlags = PostHogOnFeatureFlags { print("feature flags loaded") }
-                propertiesSanitizer =
-                    PostHogPropertiesSanitizer { properties ->
-                        properties.apply {
+//                sessionReplay = true
+//                preloadFeatureFlags = true
+//                onFeatureFlags = PostHogOnFeatureFlags { print("feature flags loaded") }
+//                propertiesSanitizer =
+//                    PostHogPropertiesSanitizer { properties ->
+//                        properties.apply {
 //                    remove("\$device_name")
-                        }
-                    }
-                sessionReplayConfig.maskAllTextInputs = false
-                sessionReplayConfig.maskAllImages = false
-                sessionReplayConfig.captureLogcat = true
-                sessionReplayConfig.screenshot = true
+//                        }
+//                    }
+//                sessionReplayConfig.maskAllTextInputs = false
+//                sessionReplayConfig.maskAllImages = false
+//                sessionReplayConfig.captureLogcat = true
+//                sessionReplayConfig.screenshot = true
             }
         PostHogAndroid.setup(this, config)
     }